pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "discord-py>=2.5.2",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.27.1",
]
